backend/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
distlib==0.3.8
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
pipenv==2023.12.1
platformdirs==4.2.0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.27
SQLAlchemy-serializer==1.4.1
typing_extensions==4.9.0
virtualenv==20.25.1
Werkzeug==3.0.1
